+++ sonolink/rest/http/player.py
@@ -37,8 +37,10 @@ class PlayerHTTPMixin:
     ) -> player.UpdatePlayerResponse:
         url = f"/sessions/{session_id}/players/{guild_id}"
         params: dict[str, str] = {}
+
         if no_replace is not None:
             params["noReplace"] = str(no_replace).lower()
+
         res = await self.request(
             "PATCH",
             url,